# Faster image rebuilding when only model artifacts are updated