--- a/bentoml/saved_bundle/bentoml-init.sh
+++ b/bentoml/saved_bundle/bentoml-init.sh
@@ -42,6 +42,7 @@ if command -v conda >/dev/null 2>&1; then
 
   echo "Updating conda base environment with environment.yml"
   conda env update -n base -f ./environment.yml
+  conda clean --all
 else
   echo "WARNING: conda command not found, skipping conda dependencies in environment.yml"
 fi